Move changelog check ver. from build systems to CI
This commit moves the changelog version check to the CI instead of the build
systems. Rationale is: the failure to update the changelog is not a build
failure / issue but rather an integration failure and it should be detected at
integration testing. The closest to integration testing is the CI testing which
is mandatory before pull requests and new features integration.

Additionally, the old version.py script is removed because it is relying on the
old version of include/tinyalsa/version.h where the TINYALSA_VERSION_STRING
macro is defiend as an explicit string literal. Since now the version string is
defined piece-wise and based on the individual version numbers, and since the
introduction of version.sh and its use within meson build system to get the
version from the version file, the old version.py script is obsolete.